services/user-service/requirements.txt
djangorestframework_simplejwt==5.5.1

